Initial commit: Support for H, W, L parameters
The goal is to be able to use H, W and L parameters in CRONs.

H-Parameter: Returns a random, but deterministic value in the field range;
if used with an incrementer, it has to be smaller than this one as well.

W: Only allowed in day of month field:
Nearest weekday to specified day. Not jumping across months. Not allowed with ranges or lists. Can be combined with 'L' to address the last weekday of the month.

L: Only allowed in day of month or day of week.
By itself, it means the last day of the week/month (SUN / 30th, 31st, 28th...).
Can be used as 'L-**day**' in day of month to address a day **day**s before the end of the month.
Can be used as '**day**L' in day of week to address the last week**day** (Mon, Tue, ...) of the month.

Added future tests for H, and some comments with ideas for L and W implementation.