New functions and tests for H parameter
H can now be used in a field by itself or with an iterator. The seed should be set before trying to parse the expression, or it will throw an error.

It will not work currently if used along "THU" in the day of week field.
Tests for a custom hash function still need to be added.