'H' now working with THU and multiple times in field, more tests
However, if H is specified multiple times in the same field, all occurrences will be replaced with the same value.
Thus, only H ranges will be useless; H in a range may produce invalid crons.
H as the iterator (and not the "offset") may also produce unwanted results with crons that trigger seldomly.

Some Tests for H added, reformatted some