do_next: 1 bit instead of byte per field
In function trying to find next cron trigger date:
Since there are only 7 fields, each field maps to a bit in an int8_t (resets) now.
If set, the field doesn't need to be reset (previously: resets byte is -1),
if bit is 0, it will be reset (prev. resets byte is field pos/not -1)