'L' offset: If too big, use first of set month; no ranges with 'H'
The offset will now be set to 30 if bigger than 30 instead of causing an error.
If an offset causes jumping into the previous month (due to lacking days, e.g. February),
it will simply go to the first of the current month.
Also, 'H' will not be able to be used in ranges, since the generated value could be below the lower or above the higher value of the range.
h_finder returns the (unchanged) field and not NULL if an error happens to allow for proper freeing of the cron expression.