'H' with selectable range
The range in which 'H' for a field will be selected can now be specified
in the form "H(min-max)".
Can still be combined with 'L' and 'W',
just not as an offset for 'L' in day-of-month field.
Simplified switch case in check_and_replace_h:
Now just choosing the max possible value for field,
to be compared with a custom max value if set or going into the replace_hashed function.