Check break cond every loop of do_next
Ensure iteration is broken when going more than 4 years in the future even when failing to search the right second (or any component in general).
On Windows gmtime seems to break eventually and return an error,
but not on macOS, so the check for a cron trigger date being too far in the future has been moved up to the start of the date finding loop.
Restructured test invalid bits:
Only 1 init for dateinit (since it can stay the same for every subtest)
Added test for completely empty cron
Fixed expr.months indexes typo
Also added a test for H with a negative range start being invalid.