'LW' compatible with lists
LW is now handled by handle_w_flags:
If the month rolls over while finding the next trigger date,
the LW flag is checked, and if present,
the next weekday will be found starting from the last day before the rolled over month.
L-Flags will now be cleared if LW is present (L can't be present in DOW then anyway) to ensure handling by w-function,
and the DOM field parsing will be skipped if it's empty because it only consisted of 'LW',
which is removed during W parsing.
TODO: Ensure only W flag is set for days with W

'LW' compatible with lists

LW is now handled by handle_w_flags:
If the month rolls over while finding the next trigger date,
the LW flag is checked, and if present,
the next weekday will be found starting from the last day before the rolled over month.
L-Flags will now be cleared if LW is present (L can't be present in DOW then anyway) to ensure handling by w-function,
and the DOM field parsing will be skipped if it's empty because it only consisted of 'LW',
which is removed during W parsing.
TODO: Ensure only W flag is set for days with W