Simplify break in find_next
If no set bit was found in the whole range from 0 to max (after no bit was found from value to max),
the range must be empty, and something in the cron_expr must be wrong.
A valid trigger value can't be found for this field then, so it makes sense to break the loop immediately instead of breaking after advancing 4 years without a trigger date.