Implement support for multiple 'H' ranges in one field
If no range is specified, they will return the same number.
But if different ranges (optionally with different offsets) are specified, different numbers should be returned due to the different modulo and/or offsets of the ranges.
Thus, multiple 'H' with ranges in one field should be allowed for convenience.