Fix replacement of 'H' in lists
The issue was the missed `free()` of the result fields being merged into the final field.
As the original subfields are already freed in `replace_hashed()`, this should now happen in the final free for the list subfields.
Also removed the `str_append()` utility, as allocating one big string and filling it with the single results (via sprintf) may take more runtime (for looping twice), but simplifies memory management and improves readability.