Redo 'L' and 'W' parsing
Instead of looping the day directly in the function,
use functions to find the days fitting for the current month,
regarding 'L' and 'W' flags.
This allows us to set multiple 'L' flags in DOM an DOW, but it may require more (do_next) loops for calculating the next date, as find_next_day may only advance to the next month at most.