test: bug for devices on new years eve
devices with the cron "0 0 12 1W,2W * ?" had problems on new years eve, triggering too often when they should only trigger once each day.
This is a cronlib issue, as it seems to jump back in time to calculate the next W flag, but doesn't realize it went behind the original start point.